Move CLI ancestor directory helpers from shared path utils
