+++ src/cli/tests/shared-ancestor-directories.test.js
@@ -2,7 +2,39 @@ import assert from "node:assert/strict";
 import path from "node:path";
 import test from "node:test";
 
-import { collectAncestorDirectories } from "../shared/dependencies.js";
+import {
+    collectAncestorDirectories,
+    collectUniqueAncestorDirectories
+} from "../shared/dependencies.js";
+
+test("collectUniqueAncestorDirectories deduplicates ancestors across inputs", () => {
+    const base = path.join(process.cwd(), "tmp", "cli-path-utils", "unique");
+    const first = path.join(base, "project", "src");
+    const second = path.join(base, "project", "tests");
+
+    const result = collectUniqueAncestorDirectories([first, second]);
+
+    const resolvedFirst = path.resolve(first);
+    const resolvedSecond = path.resolve(path.join(base, "project"));
+    const expectedRoot = path.parse(resolvedFirst).root;
+
+    assert.strictEqual(result[0], resolvedFirst);
+    assert.strictEqual(result.includes(resolvedSecond), true);
+    assert.strictEqual(result.includes(expectedRoot), true);
+});
+
+test("collectUniqueAncestorDirectories accepts a bare string", () => {
+    const base = path.join(process.cwd(), "tmp", "cli-path-utils", "string");
+    const nested = path.join(base, "src", "index.gml");
+
+    const result = collectUniqueAncestorDirectories(nested);
+
+    const resolved = path.resolve(nested);
+    const expectedRoot = path.parse(resolved).root;
+
+    assert.strictEqual(result[0], resolved);
+    assert.strictEqual(result.includes(expectedRoot), true);
+});
 
 test("collectAncestorDirectories orders and deduplicates ancestors", () => {
     const projectRoot = path.join(process.cwd(), "tmp", "cli-path-utils");
