+++ src/shared/utils/path.js
@@ -1,6 +1,5 @@
 import path from "node:path";
 
-import { toArrayFromIterable } from "./array.js";
 import { isNonEmptyString } from "./string.js";
 
 const WINDOWS_SEPARATOR_PATTERN = /\\+/g;
@@ -122,76 +121,7 @@ export function* walkAncestorDirectories(
     }
 }
 
-/**
- * Collect the unique ancestor directories for the provided starting
- * directories. Ancestors are returned in the order they were discovered so
- * callers can maintain deterministic search paths when probing for
- * configuration files.
- *
- * @param {Iterable<string | null | undefined>} startingDirectories Starting
- *        directories whose ancestors should be collected.
- * @param {{ includeSelf?: boolean }} [options]
- * @returns {Array<string>} Ordered list of unique ancestor directories.
- */
-export function collectUniqueAncestorDirectories(
-    startingDirectories,
-    { includeSelf = true } = {}
-) {
-    const directories = new Set();
-    const entries =
-        typeof startingDirectories === "string"
-            ? [startingDirectories]
-            : toArrayFromIterable(startingDirectories);
-
-    for (const start of entries) {
-        if (!isNonEmptyString(start)) {
-            continue;
-        }
-
-        for (const directory of walkAncestorDirectories(start, {
-            includeSelf
-        })) {
-            directories.add(directory);
-        }
-    }
-
-    return Array.from(directories);
-}
-
-/**
- * Checks whether {@link childPath} resides within {@link parentPath} when both
- * paths are resolved to absolute locations. Empty strings short-circuit to
- * `false` so callers can safely pass optional metadata without normalizing
- * first.
- *
- * A relative result of `""` indicates that the paths point to the same
- * directory, which is considered "inside" for consumers that treat the parent
- * as an allowed root.
- *
- * @param {string | undefined | null} childPath Path that may sit beneath
- *                                              {@link parentPath}.
- * @param {string | undefined | null} parentPath Candidate ancestor directory.
- * @returns {boolean} `true` when {@link childPath} resolves to {@link parentPath}
- *                    or a descendant.
- */
 export function isPathInside(childPath, parentPath) {
     const relative = resolveContainedRelativePath(childPath, parentPath);
     return relative !== null;
-}
-
-/**
- * Resolves every directory from the provided start paths up to the file system
- * root, preserving discovery order. Duplicate directories are returned only
- * once even when multiple starting points share ancestors. Empty inputs are
- * ignored, mirroring the truthiness guard in {@link isPathInside}.
- *
- * @param {...(string | undefined | null)} startingDirectories Path(s) whose
- *                                                             ancestor chains
- *                                                             should be
- *                                                             collected.
- * @returns {Array<string>} Flat list of absolute directories, ordered from
- *                          each start path toward the root.
- */
-export function collectAncestorDirectories(...startingDirectories) {
-    return collectUniqueAncestorDirectories(startingDirectories);
 }
